+++ src/youwol/backends/assets/routers/__init__.py
@@ -1,4 +1,5 @@
 # relative
+from .access import router as router_access
 from .files import router as router_files
 from .images import router as router_images
 from .permissions import router as router_permissions
